renamed expectation propagation to graphical
