--- autofit/expectation_propagation/factor_graphs/__init__.py
@@ -1,5 +1,0 @@
-from autofit.expectation_propagation.factor_graphs.abstract import AbstractNode
-from autofit.expectation_propagation.factor_graphs.factor import FactorValue, Factor, AbstractFactor
-from autofit.expectation_propagation.factor_graphs.graph import DeterministicFactorNode, FactorGraph
-from autofit.expectation_propagation.factor_graphs.numerical import numerical_jacobian
-from autofit.expectation_propagation.factor_graphs.variable import Plate, Variable
