+++ autofit/graphical/factor_graphs/__init__.py
@@ -1,0 +1,5 @@
+from autofit.graphical.factor_graphs.abstract import AbstractNode
+from autofit.graphical.factor_graphs.factor import FactorValue, Factor, AbstractFactor
+from autofit.graphical.factor_graphs.graph import DeterministicFactorNode, FactorGraph
+from autofit.graphical.factor_graphs.numerical import numerical_jacobian
+from autofit.graphical.factor_graphs.variable import Plate, Variable
