+++ autofit/graphical/factor_graphs/variable.py
