+++ autofit/graphical/messages/abstract.py
