+++ test_autofit/unit/config/json_priors/fit.json
@@ -82,7 +82,7 @@
                 }
             }
         },
-        "expectation_propagation": {
+        "graphical": {
             "gaussian.model.Gaussian": {
                 "sigma": {
                     "type": "Gaussian",
