+++ test_autofit/unit/graphical/gaussian/model.py
