+++ test_autofit/unit/graphical/test_factor_graph.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pytest
 
-from autofit import expectation_propagation as mp
+from autofit import graphical as mp
 
 
 def log_sigmoid(x):
